+++ CHANGES.rst
@@ -1,6 +1,9 @@
 1.0a2 (unreleased)
 ------------------
 
+- Fix CSS Lint validation and implement new 'csslint-quiet' option.
+  [hvelarde]
+
 - Fix package distribution.
   [hvelarde]